requirements.txt
asgiref==1.1.2
attrs==17.2.0
autobahn==17.9.3
Automat==0.6.0
channels==1.1.8
constantly==15.1.0
daphne==1.3.0
Django==1.11.6
django-cors-headers==2.1.0
djangorestframework==3.7.0
hyperlink==17.3.1
incremental==17.5.0
olefile==0.44
Pillow==4.3.0
pytz==2017.2
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
zope.interface==4.4.3
